Separate visual and app layer in product types
