+++ src/productTypes/components/ProductTypeListPage/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./ProductTypeListPage";
 export * from "./ProductTypeListPage";
+export * from "./filters";
