--- src/productTypes/views/ProductTypeList/messages.ts
@@ -1,22 +1,0 @@
-import { defineMessages } from "react-intl";
-
-const messages = defineMessages({
-  configurable: {
-    defaultMessage: "Configurable",
-    description: "product type"
-  },
-  digital: {
-    defaultMessage: "Digital",
-    description: "product"
-  },
-  shippable: {
-    defaultMessage: "Shippable",
-    description: "product"
-  },
-  type: {
-    defaultMessage: "Type",
-    description: "product type is digital or physical"
-  }
-});
-
-export default messages;
